+++ backend/internal/domain/donor/donor_repo.go
@@ -9,11 +9,15 @@ import (
 )
 
 type Repository interface {
+	// Write Methods
 	CreateDonorResponse(ctx context.Context, resp *donormodel.DonorResponse) (*donormodel.DonorResponse, error)
-	GetDonorResponseByID(ctx context.Context, id string) (*donormodel.DonorResponse, error)
 	UpdateDonorResponseStatus(ctx context.Context, id string, status donormodel.DonorResponseStatus) error
-	GetRecipient(ctx context.Context, id string) (*bloodreqmodel.BloodRequestWithMatchingDonors, error)
 	DeleteDonorResponse(ctx context.Context, id string) error
+	Update(ctx context.Context, resp *donormodel.DonorResponse) error
+
+	// Read Methods
+	GetDonorResponseByID(ctx context.Context, id string) (*donormodel.DonorResponse, error)
+	GetRecipient(ctx context.Context, id string) (*bloodreqmodel.BloodRequestWithMatchingDonors, error)
 	GetDonorResponsesByRequestID(ctx context.Context, reqID string) ([]*donormodel.DonorResponse, error)
 	GetDonorResponsesByDonorID(ctx context.Context, donorID string) ([]*donormodel.DonorResponse, error)
 	ExistsByID(ctx context.Context, id string) (bool, error)
@@ -23,10 +27,4 @@ type Repository interface {
 	GetByPetID(ctx context.Context, petID string) (*donormodel.DonorResponse, error)
 	GetByPetIDs(ctx context.Context, petIDs []string, ignoreSoftDelete bool) (map[string][]*donormodel.DonorResponse, error)
 	FindNotConfirmed(ctx context.Context, cutoffTime time.Time) ([]*donormodel.DonorResponse, error)
-	//=============================================
-	Accept(ctx context.Context, id string) error
-	Reject(ctx context.Context, res *donormodel.DonorResponse) error
-	Complete(ctx context.Context, id string, amount float64) error
-	Confirm(ctx context.Context, id string, amount float64) error
-	Cancel(ctx context.Context, id string) error
 }